recognizer/requirements.txt
easyocr==1.7.1
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
